create route and link page
