+++ src/index.scss
@@ -1,0 +1,18 @@
+body {
+  margin: 0;
+  font-family: 'open sans', sans-serif;
+}
+
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    monospace;
+}
+
+a {
+  text-decoration: none;
+  color: black;
+}
+
+* {
+  box-sizing: border-box;
+}
